+++ src/common/i18n/en.js
@@ -1,0 +1,18 @@
+/*
+ * /src/common/i18n/en.js
+ */
+
+pwixI18n.i18n = {
+    ...pwixI18n.i18n,
+    ...{
+        en: {
+            // a fallback icon for english-spoken countries
+            flagIcon: '/packages/pwix_i18n/images/flag-royaume-uni.png',
+            piLanguageSelector: {
+                fr: 'French',
+                en: 'English',
+                button_title: 'Choose your preferred language'
+            }
+        }
+    }
+};