+++ src/common/i18n/en_US.js
@@ -1,0 +1,12 @@
+/*
+ * /src/common/i18n/en.js
+ */
+
+pwixI18n.i18n = {
+    ...pwixI18n.i18n,
+    ...{
+        en_US: {
+            flagIcon: '/packages/pwix_i18n/images/flag-etats-unis.png'
+        }
+    }
+};