+++ src/common/i18n/fr.js
@@ -1,0 +1,17 @@
+/*
+ * /src/common/i18n/fr.js
+ */
+
+pwixI18n.i18n = {
+    ...pwixI18n.i18n,
+    ...{
+        fr: {
+            flagIcon: '/packages/pwix_i18n/images/flag-france.png',
+            piLanguageSelector: {
+                fr: 'Français',
+                en: 'Anglais',
+                button_title: 'Choisissez votre langue préférée'
+            }
+        }
+    }
+};