+++ src/common/i18n/fr_FR.js
@@ -1,0 +1,12 @@
+/*
+ * /src/common/i18n/fr.js
+ */
+
+pwixI18n.i18n = {
+    ...pwixI18n.i18n,
+    ...{
+        fr_FR: {
+            flagIcon: '/packages/pwix_i18n/images/flag-france.png'
+        }
+    }
+};