+++ src/common/js/i18n.js
@@ -1,0 +1,12 @@
+/*
+ * pwix:i18n/src/common/js/i18n.js
+ */
+
+import '../i18n/en.js';
+import '../i18n/en_US.js';
+import '../i18n/fr.js';
+import '../i18n/fr_FR.js';
+
+pwixI18n.namespace( PWIXI18NS, pwixI18n.i18n );
+
+//console.log( pwixI18n );