+++ src/common/js/index.js
@@ -5,3 +5,7 @@
 import './config.js';
 import './constants.js';
 import './functions.js';
+import './i18n.js';
+import './store.js';
+
+pwixI18n.conf.language = pwixI18n.defaultLanguage();