+++ src/server/js/check_npms.js
@@ -5,5 +5,6 @@
 import { checkNpmVersions } from 'meteor/tmeasday:check-npm-versions';
 
 checkNpmVersions({
+    'bootstrap': '^5.2',
     'printf': '^0.6.1'
 }, 'pwix:i18n' );